[AIP-21] Move operators.bash_operator operators.bash
