+++ airflow/example_dags/test_utils.py
@@ -18,7 +18,7 @@
 # under the License.
 """Used for unit tests"""
 from airflow.models import DAG
-from airflow.operators.bash_operator import BashOperator
+from airflow.operators.bash import BashOperator
 from airflow.utils.dates import days_ago
 
 dag = DAG(dag_id='test_utils', schedule_interval=None, tags=['example'])
