+++ airflow/providers/google/cloud/example_dags/example_natural_language.py
@@ -24,7 +24,7 @@ Example Airflow DAG for Google Cloud Natural Language service
 from google.cloud.language_v1.proto.language_service_pb2 import Document
 
 from airflow import models
-from airflow.operators.bash_operator import BashOperator
+from airflow.operators.bash import BashOperator
 from airflow.providers.google.cloud.operators.natural_language import (
     CloudNaturalLanguageAnalyzeEntitiesOperator, CloudNaturalLanguageAnalyzeEntitySentimentOperator,
     CloudNaturalLanguageAnalyzeSentimentOperator, CloudNaturalLanguageClassifyTextOperator,
