+++ airflow/providers/google/cloud/example_dags/example_pubsub.py
@@ -23,7 +23,7 @@ Example Airflow DAG that uses Google PubSub services.
 import os
 
 from airflow import models
-from airflow.operators.bash_operator import BashOperator
+from airflow.operators.bash import BashOperator
 from airflow.providers.google.cloud.operators.pubsub import (
     PubSubCreateSubscriptionOperator, PubSubCreateTopicOperator, PubSubDeleteSubscriptionOperator,
     PubSubDeleteTopicOperator, PubSubPublishMessageOperator,
