+++ airflow/providers/google/cloud/example_dags/example_vision.py
@@ -35,7 +35,7 @@ This DAG relies on the following OS environment variables
 import os
 
 from airflow import models
-from airflow.operators.bash_operator import BashOperator
+from airflow.operators.bash import BashOperator
 from airflow.providers.google.cloud.operators.vision import (
     CloudVisionAddProductToProductSetOperator, CloudVisionCreateProductOperator,
     CloudVisionCreateProductSetOperator, CloudVisionCreateReferenceImageOperator,
