+++ docs/howto/operator/bash.rst
@@ -22,7 +22,7 @@
 BashOperator
 ============
 
-Use the :class:`~airflow.operators.bash_operator.BashOperator` to execute
+Use the :class:`~airflow.operators.bash.BashOperator` to execute
 commands in a `Bash <https://www.gnu.org/software/bash/>`__ shell.
 
 .. exampleinclude:: ../../../airflow/example_dags/example_bash_operator.py
