+++ docs/operators-and-hooks-ref.rst
@@ -1048,7 +1048,7 @@ These integrations allow you to perform various operations using various softwar
    * - `GNU Bash <https://www.gnu.org/software/bash/>`__
      - :doc:`How to use <howto/operator/bash>`
      -
-     - :mod:`airflow.operators.bash_operator`
+     - :mod:`airflow.operators.bash`
      - :mod:`airflow.sensors.bash`
 
    * - `Kubernetes <https://kubernetes.io/>`__
