+++ tests/dags/subdir2/test_dont_ignore_this.py
@@ -19,7 +19,7 @@
 from datetime import datetime
 
 from airflow.models import DAG
-from airflow.operators.bash_operator import BashOperator
+from airflow.operators.bash import BashOperator
 
 DEFAULT_DATE = datetime(2019, 12, 1)
