+++ tests/dags/test_clear_subdag.py
@@ -21,7 +21,7 @@
 import datetime
 
 from airflow.models import DAG
-from airflow.operators.bash_operator import BashOperator
+from airflow.operators.bash import BashOperator
 from airflow.operators.subdag_operator import SubDagOperator
