+++ tests/dags/test_example_bash_operator.py
@@ -19,7 +19,7 @@
 from datetime import timedelta
 
 from airflow.models import DAG
-from airflow.operators.bash_operator import BashOperator
+from airflow.operators.bash import BashOperator
 from airflow.operators.dummy_operator import DummyOperator
 from airflow.utils.dates import days_ago
