+++ tests/dags/test_heartbeat_failed_fast.py
@@ -19,7 +19,7 @@
 from datetime import datetime
 
 from airflow.models import DAG
-from airflow.operators.bash_operator import BashOperator
+from airflow.operators.bash import BashOperator
 
 DEFAULT_DATE = datetime(2016, 1, 1)
