+++ tests/dags/test_retry_handling_job.py
@@ -20,7 +20,7 @@
 from datetime import datetime, timedelta
 
 from airflow import DAG
-from airflow.operators.bash_operator import BashOperator
+from airflow.operators.bash import BashOperator
 
 default_args = {
     'owner': 'airflow',
