+++ tests/utils/test_dot_renderer.py
@@ -23,7 +23,7 @@ import unittest
 import mock
 
 from airflow import DAG
-from airflow.operators.bash_operator import BashOperator
+from airflow.operators.bash import BashOperator
 from airflow.operators.python_operator import PythonOperator
 from airflow.utils import dot_renderer
